Split APOIndicator into several files.
