+++ js/masters/indicators/apo.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Stock/Indicators/APOIndicator.js';
+import '../../Stock/Indicators/APO/APOIndicator.js';
