+++ js/masters/indicators/indicators-all.src.js
@@ -27,7 +27,7 @@ import '../../Stock/Indicators/ChaikinIndicator.js';
 import '../../Stock/Indicators/DEMAIndicator.js';
 import '../../Stock/Indicators/TEMAIndicator.js';
 import '../../Stock/Indicators/TRIXIndicator.js';
-import '../../Stock/Indicators/APOIndicator.js';
+import '../../Stock/Indicators/APO/APOIndicator.js';
 import '../../Stock/Indicators/IKHIndicator.js';
 import '../../Stock/Indicators/KeltnerChannelsIndicator.js';
 import '../../Stock/Indicators/MACDIndicator.js';
