+++ ts/Stock/Indicators/APO/APOPoint.d.ts
@@ -1,0 +1,35 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import APOIndicator from './APOIndicator';
+import type EMAPoint from '../EMA/EMAPoint';
+
+
+/* *
+*
+*  Class
+*
+* */
+
+declare class APOPoint extends EMAPoint {
+    public series: APOIndicator;
+}
+
+/* *
+*
+*  Default Export
+*
+* */
+
+export default APOPoint;
